\name{copals_camera}
\alias{copals_camera}
\docType{data}
\title{
Copals data
}
\description{
A Liquid Chromatography Mass Spectrometry dataset made of African copals samples.  
}
\usage{data("copals_camera")}
\format{
  A data frame with 62 observations on 1022 variables.
}

\references{
M. Perrot-Dockes et al. "A multivariate variable selection approach 
for analyzing LC-MS metabolomics data",  arXiv:1704.00076
}
\examples{
data(copals_camera)
}
\keyword{datasets}
