% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group}
\alias{group}
\title{This is a qualitative variable indicating the type of tree each row of Y is.}
\description{
This is a qualitative variable indicating the type of tree each row of Y is.
}
\references{
\url{https://arxiv.org/pdf/1704.00076.pdf}
}
\author{
Marie Perrot-Dockes \email{marie.perrot-dockes@agroparistech.fr}
}
\keyword{data}
