% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metab}
\alias{metab}
\title{This is a dataset containing the abundance of 199 metabolites from 9 seeds samples
just after germination.
The temperature of seed maturation vary between the different seeds.}
\description{
This is a dataset containing the abundance of 199 metabolites from 9 seeds samples
just after germination.
The temperature of seed maturation vary between the different seeds.
}
\author{
Marie Perrot-Dockes \email{marie.perrot-dockes@agroparistech.fr}
}
\keyword{data}
