% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitening.R
\name{whitening}
\alias{whitening}
\title{This function provides an estimation of the inverse of the square root of the covariance matrix of each line of the residuals matrix.}
\usage{
whitening(residuals, typeDep, pAR = 1, qMA = 0)
}
\arguments{
\item{residuals}{the residuals matrix of independant linear model on all the collumn of the response matrix}

\item{typeDep}{character in c("AR1", "ARMA", "nonparam") defining which type of covariable to use}

\item{pAR}{numerical, only use if typeDep = "ARMA", the parameters p for the ARMA(p, q) process}
}
\value{
qMA numerical, only use if typeDep = "ARMA", the parameters p for the ARMA(p, q) process

It returns the estimation of the inverse of the square root of the covariance matrix of each line of residuals.
}
\description{
This function provides an estimation of the inverse of the square root of the covariance matrix of each line of the residuals matrix.
}
\examples{
data(copals_camera)
Y=scale(Y[,1:100])
residuals=lm(as.matrix(Y)~X-1)$residuals
whitening(residuals, "AR1")
}
