% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y}
\alias{Y}
\title{This is a metabolomic dataset from 30 copals samples of trees coming from Africa}
\description{
This is a metabolomic dataset from 30 copals samples of trees coming from Africa
}
\references{
\url{https://arxiv.org/pdf/1704.00076.pdf}
}
\author{
Marie Perrot-Dockes \email{marie.perrot-dockes@agroparistech.fr}
}
\keyword{data}
