\name{copals_camera}
\alias{copals_camera}
\docType{data}
\title{
Copals data
}
\description{
A Liquid Chromatography Mass Spectrometry dataset made of African copals samples.  
}
\usage{data("copals_camera")}
\format{
It containes Y a data frame with 30 observations on 1019 variables  and group a
qualitative variable indicating the type of tree each row of Y is.
}

\references{
M. Perrot-Dockes et al. "A multivariate variable selection approach 
for analyzing LC-MS metabolomics data",  arXiv:1704.00076
\url{https://arxiv.org/pdf/1704.00076.pdf}
}
\examples{
data(copals_camera)
}
\keyword{datasets}
